backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-multipart
passlib[argon2]
python-jose[cryptography]
pydantic-settings
openai
qdrant-client==1.9.1
httpx==0.27.2
python-dotenv==1.0.1
redis
asyncpg
pytest
pytest-asyncio
email-validator
PyMuPDF
pytest-cov
sentry-sdk[fastapi]
tenacity